{{ cookiecutter.bot_id|replace(' ', '_') }}/requirements.txt
botcity-framework-core>=0.1.1
botcity-maestro-sdk
